requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.5
Django==4.2.23
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-timezone-field==7.1
djangorestframework==3.16.0
idna==3.10
importlib_metadata==8.7.0
jwt==1.4.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
python-crontab==3.3.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.23.0
